+++ resources/lang/en/admin/planet.php
@@ -1,6 +1,14 @@
 <?php
 
 return [
+    'active' => [
+        'active' => 'Active',
+        'name' => 'Name',
+        'points' => 'Points',
+        'position' => 'Position',
+        'title' => 'Active planets',
+        'total' => 'Active planets',
+    ],
     'index' => [
         'id' => 'ID',
         'galaxy' => 'Galaxy',