+++ routes/admin.php
@@ -30,6 +30,8 @@ Route::group(['middleware' => 'auth.level:2'], function () {
 });
 
 Route::group(['middleware' => 'auth.level:1'], function () {
+    Route::get('activeplanets', 'PlanetController@active')->name('admin.planets.active');
+
     Route::get('ban', 'BanController@add')->name('admin.ban');
     Route::post('ban', 'BanController@addPost');