+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Generally speaking, with the rewrite should not come any functional changes but
 where it was unavoidable or impractical to avoid the list item denotes most notable
 changes made when rewriting.
 
+- Admiin: Active planets
+    - Translated to English
 - Admin: Ban
     - Translated to English
 - Admin: Give Moon